.github: Switch 8xlarge to 4xlarge instance_type
Switches the linux.8xlarge.nvidia.gpu to the 4xlarge instance type to
help with queueing / capacity issues. This change is only meant to be a
bridge until everyone updates their PRs to use the new
linux.4xlarge.nvidia.gpu node type

NOTE: This node type will be removed so do not depend on it for any new
workflows.

Signed-off-by: Eli Uriegas <eliuriegasfb.com>

ghstack-source-id: 724d2ce5d3fd163e235abc2066f7e09150c0db87
Pull Request resolved: https://github.com/pytorch/pytorch/pull/67299