[PyTorch] Avoid reading file from stream for 0 byte Tensor storage
Pull Request resolved: https://github.com/pytorch/pytorch/pull/67787

First noticed in https://fb.workplace.com/groups/pytorch.edge.team/posts/952737705280969/ - basically one of the speech models has ~400 0 byte tensor files, so we're basically paying the cost of looking it up in the archive and reading nothing from it.

Turns out that there's a fairly simple fix to avoid reading a 0 byte tensor. Once we notice that it's 0 bytes, just use the default `DataPtr` instead to initializing it with 0 bytes read in from the input file stream.

#accept2ship
ghstack-source-id: 142924096

Differential Revision: [D32054983](https://our.internmc.facebook.com/intern/diff/D32054983/)