dbr quantization: inline scale and zp
Summary:

Before this PR, DBR quantization used to recalculate scale and zero_point
in the converted model every time it was needed, which is slow.
This PR creates a pass during the convert function to go through every
observer in the model and cache its scale and zero_point.

Note: only doing this for observers which correspond to int8 operations
is saved for a future PR.

Test plan:

```
python test/test_quantization.py TestQuantizeDBR
```

ghstack-source-id: 72a4cff353bb69ebf8b061ac8791ee3b54cedf7e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/68251